closes #42 made some adjustemens and got rough estamates of the diameters and probobilitys.

+++ database/migrations/2020_01_12_221235_create_star_types_table.php
@@ -16,7 +16,7 @@ class CreateStarTypesTable extends Migration
         Schema::create('star_types', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('type', 32);
-            $table->mediumInteger('diameter');
+            $table->bigInteger('diameter');
             $table->string('color', 20);
             $table->tinyInteger('probability');
             $table->timestamps();
